+++ cmd/promotionerAPI/models/promotion.go
@@ -3,17 +3,17 @@ package models
 import "time"
 
 type PromotionType struct {
-	PromotionTypeID int
+	PromotionTypeID int64
 	Name            string
 }
 
 type Promotion struct {
-	PromotionID int
+	PromotionID int64
 	StartDate   time.Time
 	EndDate     time.Time
 	Text        string
 	Banner      string
 	// foreign keys
-	PromotionTypeID int
-	FoodPlaceID     int
+	PromotionTypeID int64
+	FoodPlaceID     int64
 }
